Настройка многомодульности. Добавление модуля клиента. Настройка pom.
