+++ stats-service/dto/src/main/java/ru/practicum/dto/ApplicationDtoForHitsInt.java
@@ -1,0 +1,14 @@
+package ru.practicum.dto;
+
+import lombok.*;
+
+@Getter
+@Setter
+@ToString
+@RequiredArgsConstructor
+@AllArgsConstructor
+public class ApplicationDtoForHitsInt {
+    private String app;
+    private String uri;
+    private int hits;
+}
