+++ stats-service/dto/src/main/java/ru/practicum/dto/HitDto.java
@@ -1,0 +1,15 @@
+package ru.practicum.dto;
+
+import lombok.*;
+
+@Getter
+@Setter
+@ToString
+@RequiredArgsConstructor
+@AllArgsConstructor
+public class HitDto {
+    private String app;
+    private String uri;
+    private String ip;
+    private String timestamp;
+}
