+++ stats-service/server/src/main/java/ru/practicum/Server.java
@@ -1,0 +1,13 @@
+package ru.practicum;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.annotation.PropertySource;
+
+@SpringBootApplication
+@PropertySource("classpath:application.properties")
+public class Server {
+    public static void main(String[] args) {
+        SpringApplication.run(Server.class, args);
+    }
+}
