+++ stats-service/server/src/main/java/ru/practicum/dtoSer/ApplicationDto.java
@@ -1,0 +1,12 @@
+package ru.practicum.dtoSer;
+
+import lombok.*;
+
+@Getter
+@Setter
+@ToString
+@RequiredArgsConstructor
+@AllArgsConstructor
+public class ApplicationDto {
+    private String app;
+}
