+++ stats-service/server/src/main/java/ru/practicum/exceptions/ErrorResponse.java
@@ -1,0 +1,12 @@
+package ru.practicum.exceptions;
+
+import lombok.Getter;
+
+@Getter
+public class ErrorResponse {
+    private final String error;
+
+    public ErrorResponse(String error) {
+        this.error = error;
+    }
+}
