+++ stats-service/server/src/main/java/ru/practicum/service/HitService.java
@@ -1,0 +1,14 @@
+package ru.practicum.service;
+
+import org.springframework.http.HttpStatus;
+import ru.practicum.dto.ApplicationDtoForHitsInt;
+import ru.practicum.dto.HitDto;
+
+import java.time.LocalDateTime;
+import java.util.List;
+
+public interface HitService {
+    HttpStatus addHit(HitDto hitDto);
+
+    List<ApplicationDtoForHitsInt> getStats(LocalDateTime start, LocalDateTime end, List<String> uris, boolean unique);
+}
