+++ stats-service/server/src/main/resources/application.properties
