+++ stats-service/server/src/main/resources/schema.sql
