--- stats/src/main/java/ru/practicum/stat/client/controller/StatClient.java
@@ -1,4 +1,0 @@
-package ru.practicum.stat.client.controller;
-
-public class StatClient {
-}
