--- stats/src/main/java/ru/practicum/stat/service/dto/ApplicationDto.java
@@ -1,12 +1,0 @@
-package ru.practicum.stat.service.dto;
-
-import lombok.*;
-
-@Getter
-@Setter
-@ToString
-@RequiredArgsConstructor
-@AllArgsConstructor
-public class ApplicationDto {
-    private String app;
-}
