--- stats/src/main/java/ru/practicum/stat/service/dto/ApplicationDtoForHits.java
@@ -1,14 +1,0 @@
-package ru.practicum.stat.service.dto;
-
-import lombok.*;
-
-@Getter
-@Setter
-@ToString
-@RequiredArgsConstructor
-@AllArgsConstructor
-public class ApplicationDtoForHits {
-    private String app;
-    private String uri;
-    private Long hits;
-}
