--- stats/src/main/java/ru/practicum/stat/service/dto/ApplicationDtoForHitsInt.java
@@ -1,14 +1,0 @@
-package ru.practicum.stat.service.dto;
-
-import lombok.*;
-
-@Getter
-@Setter
-@ToString
-@RequiredArgsConstructor
-@AllArgsConstructor
-public class ApplicationDtoForHitsInt {
-    private String app;
-    private String uri;
-    private int hits;
-}
