--- stats/src/main/java/ru/practicum/stat/service/dto/HitDto.java
@@ -1,15 +1,0 @@
-package ru.practicum.stat.service.dto;
-
-import lombok.*;
-
-@Getter
-@Setter
-@ToString
-@RequiredArgsConstructor
-@AllArgsConstructor
-public class HitDto {
-    private String app;
-    private String uri;
-    private String ip;
-    private String timestamp;
-}
