--- stats/src/main/java/ru/practicum/stat/service/exceptions/DataTimeException.java
@@ -1,10 +1,0 @@
-package ru.practicum.stat.service.exceptions;
-
-import lombok.Getter;
-
-@Getter
-public class DataTimeException extends RuntimeException {
-    public DataTimeException(String message) {
-        super(message);
-    }
-}
