--- stats/src/main/java/ru/practicum/stat/service/exceptions/ErrorHandler.java
@@ -1,20 +1,0 @@
-package ru.practicum.stat.service.exceptions;
-
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.ResponseStatus;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-
-@RestControllerAdvice
-@Slf4j
-public class ErrorHandler {
-
-    @ExceptionHandler
-    @ResponseStatus(HttpStatus.BAD_REQUEST)
-    public ErrorResponse errorBadRequest(final DataTimeException e) {
-        return new ErrorResponse(
-                e.getMessage()
-        );
-    }
-}
