--- stats/src/main/java/ru/practicum/stat/service/exceptions/ErrorResponse.java
@@ -1,12 +1,0 @@
-package ru.practicum.stat.service.exceptions;
-
-import lombok.Getter;
-
-@Getter
-public class ErrorResponse {
-    private final String error;
-
-    public ErrorResponse(String error) {
-        this.error = error;
-    }
-}
