--- stats/src/main/java/ru/practicum/stat/service/repository/ApplicationRepository.java
@@ -1,12 +1,0 @@
-package ru.practicum.stat.service.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-import ru.practicum.stat.service.model.Application;
-
-import java.util.Optional;
-
-@Repository
-public interface ApplicationRepository extends JpaRepository<Application, Long> {
-    Optional<Application> findByApp(String app);
-}
